+++ resources/views/emails/insight/data-dive.blade.php
@@ -1,0 +1,2 @@
+@php($productName = 'Data Dive')
+@include('emails.insight._product-issue')