+++ resources/views/emails/insight/marina-maitama.blade.php
@@ -1,0 +1,2 @@
+@php($productName = 'Marina and Maitama')
+@include('emails.insight._product-issue')