+++ resources/views/emails/insight/pocket-science.blade.php
@@ -1,0 +1,2 @@
+@php($productName = 'Pocket Science')
+@include('emails.insight._product-issue')