+++ resources/views/emails/insight/senorrita.blade.php
@@ -1,0 +1,2 @@
+@php($productName = 'SenorRita')
+@include('emails.insight._product-issue')